fix cis2 tokens duplication
